\name{solid-package}
\alias{solid-package}
\alias{solid}
\docType{package}
\title{
A Divide-and-Conquer Method for Sparse Risk Prediction and Evaluation
}
\description{
The package \pkg{solid} consists 1) a screening and one-step linearization infused DAC (SOLID) algorithm to fit sparse logistic regression to massive datasets, and 2) a modified cross-validation (MCV) that utilizes the side products of the SOLID hence substantially reduce the computational burden.
}
\details{
\tabular{ll}{
Package: \tab solid\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2019-11-06\cr
License: GPL>=2 \cr
}
}

\author{
Chuan Hong \cr
Yan Wang \cr
Tianxi Cai \cr
Maintainer: Chuan Hong <chuan_hong@hms.harvard.edu>
}

\keyword{package}

